requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
tensorflow>=2.8.0
keras>=2.8.0

# XAI Libraries
shap>=0.41.0
lime>=0.2.0.1
alibi>=0.7.0
captum>=0.5.0
interpret>=0.2.7
anchor-exp>=0.0.1

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0
albumentations>=1.1.0

# Text Processing
nltk>=3.6.0
spacy>=3.4.0
textblob>=0.15.3

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.4.0

# Configuration and Utilities
pyyaml>=6.0
click>=8.0.0
tqdm>=4.62.0
joblib>=1.1.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipywidgets>=7.6.0

# Additional ML Libraries
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Evaluation and Metrics
scikit-image>=0.19.0
scikit-multiflow>=0.5.0

# Statistical Analysis
statsmodels>=0.13.0

# Dataset Loading
datasets>=2.0.0

# Utilities
requests>=2.28.0
tqdm>=4.64.0
colorama>=0.4.4 